[update] added sidebar
